Add basic doc
Add some todos
